requirements.txt
Flask
mysqlclient
social-auth-app-flask
flask-login
social-auth-app-flask-sqlalchemy
flask-sqlalchemy